chore: pull out log config into separate module
