+++ eel_hole/__init__.py
@@ -1,13 +1,11 @@
 """Main app definition."""
 
 import json
-import logging
 import os
 from dataclasses import asdict
 from urllib.parse import quote
 
 import requests
-import structlog
 from authlib.integrations.flask_client import OAuth
 from flask import (
     Flask,
@@ -31,34 +29,13 @@ from frictionless import Package
 
 from eel_hole.models import db, User
 from eel_hole.duckdb_query import ag_grid_to_duckdb, Filter
+from eel_hole.logs import log
 from eel_hole.search import initialize_index, run_search
 from eel_hole.utils import clean_descriptions
 
 AUTH0_DOMAIN = os.getenv("PUDL_VIEWER_AUTH0_DOMAIN")
 CLIENT_ID = os.getenv("PUDL_VIEWER_AUTH0_CLIENT_ID")
 CLIENT_SECRET = os.getenv("PUDL_VIEWER_AUTH0_CLIENT_SECRET")
-
-
-def user_id_adder(logger, log_method, event_dict):
-    """Add user ID to log if available.
-
-    Must be added to processor list *before* JSONRenderer, otherwise event_dict
-    will be rendered to string already.
-    """
-    if current_user:
-        event_dict["user_id"] = current_user.get_id()
-    return event_dict
-
-
-structlog.configure(
-    processors=[
-        structlog.processors.TimeStamper(fmt="iso"),
-        user_id_adder,
-        structlog.processors.JSONRenderer(),
-    ],
-    wrapper_class=structlog.make_filtering_bound_logger(logging.INFO),
-)
-log = structlog.get_logger()
 
 
 def __init_auth0(app: Flask):
